refactor: change how library is imported
BREAKING CHANGE: `@nuxtjs/sanity/module` is now how the module should be imported in `nuxt.config`
