+++ src/components/sanity-content.ts
@@ -2,8 +2,8 @@ import defu from 'defu'
 import type { CreateElement, VNode, VNodeData, VueConstructor } from 'vue'
 
 // eslint-disable-next-line
-import type { SanityConfiguration } from '..'
-import { extendVue } from '../vue'
+import type { SanityConfiguration } from '../module'
+import { extendVue } from './vue'
 
 interface MarkDefs {
   _key: string
